requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiomysql==0.2.0
aiosignal==1.3.1
alpha_vantage==3.0.0
annotated-types==0.7.0
anyio==4.6.2.post1
appdirs==1.4.4
attrs==24.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
cachetools==5.5.1
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
cssselect==1.2.0
deep-translator==1.11.4
ephem==4.1.6
fake-useragent==2.0.3
feedparser==6.0.11
filelock==3.17.0
frozendict==2.4.6
frozenlist==1.5.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.66.0
grpcio==1.69.0
grpcio-status==1.69.0
h11==0.14.0
html2text==2024.2.26
html5lib==1.1
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
idna==3.10
IMAPClient==3.0.1
importlib_metadata==8.5.0
LunarCalendar==0.0.9
lxml==5.3.0
multidict==6.1.0
multitasking==0.0.11
mysql-connector-python==9.1.0
numpy==2.2.1
pandas==2.2.3
pandas-datareader==0.10.0
parse==1.20.2
peewee==3.17.8
platformdirs==4.3.6
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.29.3
psutil==6.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.5
pydantic_core==2.27.2
pyee==11.1.1
PyMySQL==1.1.1
pyparsing==3.2.1
pyppeteer==2.0.0
pyquery==2.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.11.1
pytz==2024.2
requests==2.32.3
requests-html==0.10.0
rsa==4.9
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
tencentcloud-sdk-python==3.0.1266
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
uritools==4.0.3
urlextract==1.9.0
urllib3==1.26.20
w3lib==2.2.1
webencodings==0.5.1
websockets==10.4
yahoo-fin==0.8.9.1
yarl==1.16.0
yfinance==0.2.51
zipp==3.21.0
